candid
Struct Principal

pub struct Principal { /* private fields */ }

Generic ID on Internet Computer.

Principals are generic identifiers for canisters, users and possibly other concepts in the future. As far as most uses of the IC are concerned they are opaque binary blobs with a length between 0 and 29 bytes, and there is intentionally no mechanism to tell canister ids and user ids apart.

Note a principal is not necessarily tied with a public key-pair, yet we need at least a key-pair of a related principal to sign requests.

A Principal can be serialized to a byte array (Vec<u8>) or a text representation, but the inner structure of the byte representation is kept private.

Example of using a Principal object:

use ic_principal::Principal;

let text = "aaaaa-aa";  // The management canister ID.
let principal = Principal::from_text(text).expect("Could not decode the principal.");
assert_eq!(principal.as_slice(), &[]);
assert_eq!(principal.to_text(), text);

Serialization is enabled with the “serde” feature. It supports serializing to a byte bufer for non-human readable serializer, and a string version for human readable serializers.

use ic_principal::Principal;
use serde::{Deserialize, Serialize};
use std::str::FromStr;

#[derive(Serialize)]
struct Data {
    id: Principal,
}

let id = Principal::from_str("2chl6-4hpzw-vqaaa-aaaaa-c").unwrap();

// JSON is human readable, so this will serialize to a textual
// representation of the Principal.
assert_eq!(
    serde_json::to_string(&Data { id: id.clone() }).unwrap(),
    r#"{"id":"2chl6-4hpzw-vqaaa-aaaaa-c"}"#
);

// CBOR is not human readable, so will serialize to bytes.
assert_eq!(
    serde_cbor::to_vec(&Data { id: id.clone() }).unwrap(),
    &[161, 98, 105, 100, 73, 239, 205, 171, 0, 0, 0, 0, 0, 1],
);

Implementations

impl Principal

pub const MAX_LENGTH_IN_BYTES: usize = 29usize

pub const fn management_canister() -> Principal

Construct a Principal of the IC management canister

pub fn self_authenticating<P>(public_key: P) -> Principal
where P: AsRef<[u8]>,

Construct a self-authenticating ID from public key

pub const fn anonymous() -> Principal

Construct an anonymous ID.

pub const fn from_slice(slice: &[u8]) -> Principal

Construct a Principal from a slice of bytes.

Panics

Panics if the slice is longer than 29 bytes.

pub const fn try_from_slice(slice: &[u8]) -> Result<Principal, PrincipalError>

Construct a Principal from a slice of bytes.

pub fn from_text<S>(text: S) -> Result<Principal, PrincipalError>
where S: AsRef<str>,

Parse a Principal from text representation.

pub fn to_text(&self) -> String

Convert Principal to text representation.

pub fn as_slice(&self) -> &[u8]

Return the Principal’s underlying slice of bytes.