candid::types
Module reference

Data structure for Candid value Func and Service Note that Func and Service should not be used directly. We need to define a newtype for Func or Service, and manually impl CandidType for the newtype, in order to specify the correct reference type. We have two macros define_function! and define_service! to help defining the newtype.

Structs

Func

FuncVisitor

A Visitor to extract Funcs.

Service