candid
Module utils

Traits

ArgumentDecoder

Allow decoding of any sized argument.

ArgumentEncoder

Allow encoding of any serializable value.

Functions

check_unique

decode_args

Decode a series of arguments, represented as a tuple. There is a maximum of 16 arguments supported.

decode_args_with_config

decode_args_with_config_debug

decode_one

Decode a single argument.

decode_one_with_config

encode_args

Serialize an encoding of a tuple to a vector of bytes.

encode_one

Serialize a single value to a vector of bytes.

pp_num_str

write_args

Serialize an encoding of a tuple and write it to a Write buffer.